requirements.txt
django
djangorestframework
psycopg2
python-dotenv
pytest 
pytest-django
factory_boy